auth-service-api/auth/requirements.txt
# API framework
fastapi==0.111.0
uvicorn[standard]==0.30.1

python-jose[cryptography]
passlib[bcrypt]
python-multipart
